tr1/inv-ring: extract priv module
