+++ src/tr1/game/inventory_ring/priv.c
@@ -3,14 +3,13 @@
 #include "game/inventory.h"
 #include "game/output.h"
 #include "global/vars.h"
-
 #include <libtrx/game/math.h>
 
 static GAME_OBJECT_ID m_RequestedObjectID = NO_OBJECT;
 
-static void M_HandleRequestedObject(INV_RING *ring);
+static void M_HandleRequestedObject(RING_INFO *ring);
 
-static void M_HandleRequestedObject(INV_RING *const ring)
+static void M_HandleRequestedObject(RING_INFO *const ring)
 {
     if (m_RequestedObjectID == NO_OBJECT) {
         return;
@@ -33,8 +32,8 @@ void InvRing_SetRequestedObjectID(const GAME_OBJECT_ID object_id)
 }
 
 void InvRing_InitRing(
-    INV_RING *const ring, const RING_TYPE type, INVENTORY_ITEM **list,
-    const int16_t qty, const int16_t current)
+    RING_INFO *ring, int16_t type, INVENTORY_ITEM **list, int16_t qty,
+    int16_t current, IMOTION_INFO *imo)
 {
     ring->type = type;
     ring->radius = 0;
@@ -42,10 +41,6 @@ void InvRing_InitRing(
     ring->number_of_objects = qty;
     ring->current_object = current;
     ring->angle_adder = 0x10000 / qty;
-
-    ring->is_pass_open = false;
-    ring->is_demo_needed = false;
-    ring->has_spun_out = false;
 
     M_HandleRequestedObject(ring);
 
@@ -61,6 +56,8 @@ void InvRing_InitRing(
     ring->rot_adder_l = 0;
     ring->rot_adder_r = 0;
 
+    ring->imo = imo;
+
     ring->camera.pos.x = 0;
     ring->camera.pos.y = CAMERA_STARTHEIGHT;
     ring->camera.pos.z = 896;
@@ -68,19 +65,19 @@ void InvRing_InitRing(
     ring->camera.rot.y = 0;
     ring->camera.rot.z = 0;
 
-    InvRing_MotionInit(ring, RNG_OPENING, RNG_OPEN, OPEN_FRAMES);
+    InvRing_MotionInit(ring, OPEN_FRAMES, RNG_OPENING, RNG_OPEN);
     InvRing_MotionRadius(ring, RING_RADIUS);
     InvRing_MotionCameraPos(ring, CAMERA_HEIGHT);
     InvRing_MotionRotation(
         ring, OPEN_ROTATION,
         0xC000 - (ring->current_object * ring->angle_adder));
 
-    ring->ring_pos.pos.x = 0;
-    ring->ring_pos.pos.y = 0;
-    ring->ring_pos.pos.z = 0;
-    ring->ring_pos.rot.x = 0;
-    ring->ring_pos.rot.y = ring->motion.rotate_target - OPEN_ROTATION;
-    ring->ring_pos.rot.z = 0;
+    ring->ringpos.pos.x = 0;
+    ring->ringpos.pos.y = 0;
+    ring->ringpos.pos.z = 0;
+    ring->ringpos.rot.x = 0;
+    ring->ringpos.rot.y = imo->rotate_target - OPEN_ROTATION;
+    ring->ringpos.rot.z = 0;
 
     ring->light.x = -1536;
     ring->light.y = 256;
@@ -89,21 +86,21 @@ void InvRing_InitRing(
 
 void InvRing_ResetItem(INVENTORY_ITEM *const inv_item)
 {
-    inv_item->meshes_drawn = inv_item->meshes_sel;
+    inv_item->drawn_meshes = inv_item->which_meshes;
     inv_item->current_frame = 0;
     inv_item->goal_frame = inv_item->current_frame;
-    inv_item->x_rot_pt = 0;
+    inv_item->pt_xrot = 0;
     inv_item->x_rot = 0;
     inv_item->y_rot = 0;
-    inv_item->y_trans = 0;
-    inv_item->z_trans = 0;
+    inv_item->ytrans = 0;
+    inv_item->ztrans = 0;
     inv_item->action = ACTION_USE;
     if (inv_item->object_id == O_PASSPORT_OPTION) {
         inv_item->object_id = O_PASSPORT_CLOSED;
     }
 }
 
-void InvRing_GetView(INV_RING *const ring, XYZ_32 *view_pos, XYZ_16 *view_rot)
+void InvRing_GetView(RING_INFO *ring, XYZ_32 *view_pos, XYZ_16 *view_rot)
 {
     PHD_ANGLE angles[2];
 
@@ -118,7 +115,7 @@ void InvRing_GetView(INV_RING *const ring, XYZ_32 *view_pos, XYZ_16 *view_rot)
     view_rot->z = 0;
 }
 
-void InvRing_Light(INV_RING *const ring)
+void InvRing_Light(RING_INFO *ring)
 {
     PHD_ANGLE angles[2];
     Output_SetLightDivider(0x6000);
@@ -126,75 +123,187 @@ void InvRing_Light(INV_RING *const ring)
     Output_RotateLight(angles[1], angles[0]);
 }
 
-void InvRing_CalcAdders(INV_RING *const ring, int16_t rotation_duration)
+void InvRing_CalcAdders(RING_INFO *ring, int16_t rotation_duration)
 {
     ring->angle_adder = 0x10000 / ring->number_of_objects;
     ring->rot_adder_l = ring->angle_adder / rotation_duration;
     ring->rot_adder_r = -ring->rot_adder_l;
 }
 
-void InvRing_DoMotions(INV_RING *const ring)
+void InvRing_DoMotions(RING_INFO *ring)
 {
-    INV_RING_MOTION *const motion = &ring->motion;
+    IMOTION_INFO *imo = ring->imo;
 
-    if (motion->count != 0) {
-        ring->radius += motion->radius_rate;
-        ring->camera.pos.y += motion->camera_y_rate;
-        ring->ring_pos.rot.y += motion->rotate_rate;
-        ring->camera_pitch += motion->camera_pitch_rate;
+    if (imo->count) {
+        ring->radius += imo->radius_rate;
+        ring->camera.pos.y += imo->camera_yrate;
+        ring->ringpos.rot.y += imo->rotate_rate;
+        ring->camera_pitch += imo->camera_pitch_rate;
 
-        INVENTORY_ITEM *const inv_item = ring->list[ring->current_object];
-        inv_item->x_rot_pt += motion->item_pt_x_rot_rate;
-        inv_item->x_rot += motion->item_x_rot_rate;
-        inv_item->y_trans += motion->item_y_trans_rate;
-        inv_item->z_trans += motion->item_z_trans_rate;
+        INVENTORY_ITEM *inv_item = ring->list[ring->current_object];
+        inv_item->pt_xrot += imo->item_ptxrot_rate;
+        inv_item->x_rot += imo->item_xrot_rate;
+        inv_item->ytrans += imo->item_ytrans_rate;
+        inv_item->ztrans += imo->item_ztrans_rate;
 
-        motion->count--;
-        if (!motion->count) {
-            motion->status = motion->status_target;
-            if (motion->radius_rate) {
-                motion->radius_rate = 0;
-                ring->radius = motion->radius_target;
+        imo->count--;
+        if (!imo->count) {
+            imo->status = imo->status_target;
+            if (imo->radius_rate) {
+                imo->radius_rate = 0;
+                ring->radius = imo->radius_target;
             }
-            if (motion->camera_y_rate) {
-                motion->camera_y_rate = 0;
-                ring->camera.pos.y = motion->camera_y_target;
+            if (imo->camera_yrate) {
+                imo->camera_yrate = 0;
+                ring->camera.pos.y = imo->camera_ytarget;
             }
-            if (motion->rotate_rate) {
-                motion->rotate_rate = 0;
-                ring->ring_pos.rot.y = motion->rotate_target;
+            if (imo->rotate_rate) {
+                imo->rotate_rate = 0;
+                ring->ringpos.rot.y = imo->rotate_target;
             }
-            if (motion->item_pt_x_rot_rate) {
-                motion->item_pt_x_rot_rate = 0;
-                inv_item->x_rot_pt = motion->item_pt_x_rot_target;
+            if (imo->item_ptxrot_rate) {
+                imo->item_ptxrot_rate = 0;
+                inv_item->pt_xrot = imo->item_ptxrot_target;
             }
-            if (motion->item_x_rot_rate) {
-                motion->item_x_rot_rate = 0;
-                inv_item->x_rot = motion->item_x_rot_target;
+            if (imo->item_xrot_rate) {
+                imo->item_xrot_rate = 0;
+                inv_item->x_rot = imo->item_xrot_target;
             }
-            if (motion->item_y_trans_rate) {
-                motion->item_y_trans_rate = 0;
-                inv_item->y_trans = motion->item_y_trans_target;
+            if (imo->item_ytrans_rate) {
+                imo->item_ytrans_rate = 0;
+                inv_item->ytrans = imo->item_ytrans_target;
             }
-            if (motion->item_z_trans_rate) {
-                motion->item_z_trans_rate = 0;
-                inv_item->z_trans = motion->item_z_trans_target;
+            if (imo->item_ztrans_rate) {
+                imo->item_ztrans_rate = 0;
+                inv_item->ztrans = imo->item_ztrans_target;
             }
-            if (motion->camera_pitch_rate) {
-                motion->camera_pitch_rate = 0;
-                ring->camera_pitch = motion->camera_pitch_target;
+            if (imo->camera_pitch_rate) {
+                imo->camera_pitch_rate = 0;
+                ring->camera_pitch = imo->camera_pitch_target;
             }
         }
     }
 
     if (ring->rotating) {
-        ring->ring_pos.rot.y += ring->rot_adder;
+        ring->ringpos.rot.y += ring->rot_adder;
         ring->rot_count--;
         if (!ring->rot_count) {
             ring->current_object = ring->target_object;
-            ring->ring_pos.rot.y =
+            ring->ringpos.rot.y =
                 0xC000 - (ring->current_object * ring->angle_adder);
             ring->rotating = 0;
         }
     }
+}
+
+void InvRing_RotateLeft(RING_INFO *ring)
+{
+    ring->rotating = 1;
+    ring->target_object = ring->current_object - 1;
+    if (ring->target_object < 0) {
+        ring->target_object = ring->number_of_objects - 1;
+    }
+    ring->rot_count = ROTATE_DURATION;
+    ring->rot_adder = ring->rot_adder_l;
+}
+
+void InvRing_RotateRight(RING_INFO *ring)
+{
+    ring->rotating = 1;
+    ring->target_object = ring->current_object + 1;
+    if (ring->target_object >= ring->number_of_objects) {
+        ring->target_object = 0;
+    }
+    ring->rot_count = ROTATE_DURATION;
+    ring->rot_adder = ring->rot_adder_r;
+}
+
+void InvRing_MotionInit(
+    RING_INFO *ring, int16_t frames, int16_t status, int16_t status_target)
+{
+    ring->imo->status_target = status_target;
+    ring->imo->count = frames;
+    ring->imo->status = status;
+    ring->imo->radius_target = 0;
+    ring->imo->radius_rate = 0;
+    ring->imo->camera_ytarget = 0;
+    ring->imo->camera_yrate = 0;
+    ring->imo->camera_pitch_target = 0;
+    ring->imo->camera_pitch_rate = 0;
+    ring->imo->rotate_target = 0;
+    ring->imo->rotate_rate = 0;
+    ring->imo->item_ptxrot_target = 0;
+    ring->imo->item_ptxrot_rate = 0;
+    ring->imo->item_xrot_target = 0;
+    ring->imo->item_xrot_rate = 0;
+    ring->imo->item_ytrans_target = 0;
+    ring->imo->item_ytrans_rate = 0;
+    ring->imo->item_ztrans_target = 0;
+    ring->imo->item_ztrans_rate = 0;
+    ring->imo->misc = 0;
+}
+
+void InvRing_MotionSetup(
+    RING_INFO *ring, int16_t status, int16_t status_target, int16_t frames)
+{
+    IMOTION_INFO *imo = ring->imo;
+    imo->count = frames;
+    imo->status = status;
+    imo->status_target = status_target;
+    imo->radius_rate = 0;
+    imo->camera_yrate = 0;
+}
+
+void InvRing_MotionRadius(RING_INFO *ring, int16_t target)
+{
+    IMOTION_INFO *imo = ring->imo;
+    imo->radius_target = target;
+    imo->radius_rate = (target - ring->radius) / imo->count;
+}
+
+void InvRing_MotionRotation(RING_INFO *ring, int16_t rotation, int16_t target)
+{
+    IMOTION_INFO *imo = ring->imo;
+    imo->rotate_target = target;
+    imo->rotate_rate = rotation / imo->count;
+}
+
+void InvRing_MotionCameraPos(RING_INFO *ring, int16_t target)
+{
+    IMOTION_INFO *imo = ring->imo;
+    imo->camera_ytarget = target;
+    imo->camera_yrate = (target - ring->camera.pos.y) / imo->count;
+}
+
+void InvRing_MotionCameraPitch(RING_INFO *ring, int16_t target)
+{
+    IMOTION_INFO *imo = ring->imo;
+    imo->camera_pitch_target = target;
+    imo->camera_pitch_rate = target / imo->count;
+}
+
+void InvRing_MotionItemSelect(RING_INFO *ring, INVENTORY_ITEM *inv_item)
+{
+    IMOTION_INFO *imo = ring->imo;
+    imo->item_ptxrot_target = inv_item->pt_xrot_sel;
+    imo->item_ptxrot_rate = inv_item->pt_xrot_sel / imo->count;
+    imo->item_xrot_target = inv_item->x_rot_sel;
+    imo->item_xrot_rate = inv_item->x_rot_sel / imo->count;
+    imo->item_ytrans_target = inv_item->ytrans_sel;
+    imo->item_ytrans_rate = inv_item->ytrans_sel / imo->count;
+    imo->item_ztrans_target = inv_item->ztrans_sel;
+    imo->item_ztrans_rate = inv_item->ztrans_sel / imo->count;
+}
+
+void InvRing_MotionItemDeselect(RING_INFO *ring, INVENTORY_ITEM *inv_item)
+{
+    IMOTION_INFO *imo = ring->imo;
+    imo->item_ptxrot_target = 0;
+    imo->item_ptxrot_rate = -inv_item->pt_xrot_sel / imo->count;
+    imo->item_xrot_target = 0;
+    imo->item_xrot_rate = -inv_item->x_rot_sel / imo->count;
+    imo->item_ytrans_target = 0;
+    imo->item_ytrans_rate = -inv_item->ytrans_sel / imo->count;
+    imo->item_ztrans_target = 0;
+    imo->item_ztrans_rate = -inv_item->ztrans_sel / imo->count;
 }
