+++ src/tr1/meson.build
@@ -116,6 +116,7 @@ sources = [
   'game/inventory.c',
   'game/inventory_ring/control.c',
   'game/inventory_ring/draw.c',
+  'game/inventory_ring/priv.c',
   'game/inventory_ring/vars.c',
   'game/item_actions.c',
   'game/item_actions/bubbles.c',
